refactor: delegate multicall to lib
