+++ src/state/index.ts
@@ -1,5 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/query/react'
+import multicall from 'lib/state/multicall'
 import { load, save } from 'redux-localstorage-simple'
 
 import application from './application/reducer'
@@ -11,7 +12,6 @@ import lists from './lists/reducer'
 import logs from './logs/slice'
 import mint from './mint/reducer'
 import mintV3 from './mint/v3/reducer'
-import { multicall } from './multicall/instance'
 import { routingApi } from './routing/slice'
 import swap from './swap/reducer'
 import transactions from './transactions/reducer'
