--- src/state/multicall/instance.ts
@@ -1,4 +1,0 @@
-import { createMulticall } from '@uniswap/redux-multicall'
-
-// Create a multicall instance with default settings
-export const multicall = createMulticall()
